Refactor Wasm generation
