+++ crates/cli/src/wasm_generator/mod.rs
@@ -1,0 +1,5 @@
+mod dynamic;
+pub mod r#static;
+mod transform;
+
+pub use dynamic::Generator as Dynamic;
