+++ crates/cli/src/wasm_generator/transform.rs
